# specs/001-phase-05-advanced-features/contracts/backend-api.openapi.yaml
openapi: 3.0.3
info:
  title: TaskAI Backend API
  description: Core API for TaskAI task management system with advanced features
  version: 1.0.0
  contact:
    name: TaskAI Team
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.taskai.example.com
    description: Production
security:
  - BearerAuth: []
...
paths:
...
  /api/tasks/{task_id}/tags:
    post:
      tags: [Tasks]
      summary: Add tag to task
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tag_name]
              properties:
                tag_name:
                  type: string
                  maxLength: 100
      responses:
        '200':
          description: Tag added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Maximum tags limit reached (10)

    delete:
      tags: [Tasks]
      summary: Remove tag from task
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - name: tag_name
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tag removed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
...
  /api/tags:
    get:
      tags: [Tags]
      summary: List user's tags
      responses:
        '200':
          description: Tags retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'

    post:
      tags: [Tags]
      summary: Create new tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 100
                color:
                  type: string
                  pattern: '^#[0-9A-Fa-f]{6}$'
      responses:
        '201':
          description: Tag created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          description: Maximum tags limit reached (100)
        '409':
          description: Tag name already exists (case-insensitive)
  /api/tags/{tag_id}:
    put:
      tags: [Tags]
      summary: Update tag
      parameters:
        - name: tag_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 100
                color:
                  type: string
                  pattern: '^#[0-9A-Fa-f]{6}$'
      responses:
        '200':
          description: Tag updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'

    delete:
      tags: [Tags]
      summary: Delete tag
      parameters:
        - name: tag_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Tag deleted successfully
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TaskId:
      name: task_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        completed:
          type: boolean
        priority:
          type: string
          enum: [Low, Medium, High]
        due_date:
          type: string
          format: date-time
          nullable: true
        is_recurring:
          type: boolean
        parent_task_id:
          type: string
          format: uuid
          nullable: true
        recurrence_rule:
          $ref: '#/components/schemas/RecurrenceRule'
          nullable: true
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        reminders:
          type: array
          items:
            $ref: '#/components/schemas/Reminder'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Tag:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
        usage_count:
          type: integer
        created_at:
          type: string
          format: date-time
    Reminder:
      type: object
      properties:
        id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        scheduled_time:
          type: string
          format: date-time
        reminder_type:
          type: string
          enum: [15min, 1hr, 1day, 1week, custom]
        status:
          type: string
          enum: [pending, sent, failed]
...
    RecurrenceRule:
      type: object
      properties:
        id:
          type: string
          format: uuid
        recurrence_type:
          type: string
          enum: [daily, weekly, monthly, yearly, custom]
        interval:
          type: integer
          minimum: 1
        days_of_week:
          type: array
          items:
            type: integer
            minimum: 0
            maximum: 6
          nullable: true
        day_of_month:
          type: integer
          minimum: 1
          maximum: 31
          nullable: true
        month_of_year:
          type: integer
          minimum: 1
          maximum: 12
          nullable: true
        custom_pattern:
          type: string
          nullable: true
...